minor fixes
Home view: changed loading spinner to activity indicator and made it center
authslice.js : for new users, lightning address was not getting set, so tried to update this in loading profile view but did not work.
loadingprofile.js added dispatch code to set username but did not work...!
walletinfoscreen.js: Made it as Text so that it can be copied
settingshomescreen.js : reset twittermodel indicator when logging out.